requirements.txt
sentence-transformers==3.0.1
transformers==4.44.2
faiss-gpu==1.7.2
pypdf==4.3.1
gradio==4.43.0
pandas==2.2.2 
datasets==2.21.0
python-dotenv==1.0.1
google-generativeai==0.7.2
bitsandbytes==0.43.3
accelerate==0.34.2
sentencepiece==0.2.0